Desktop/painting/requirements.txt
opencv-python>=4.8.0
torch>=2.6.0
torchvision>=0.19.0
Pillow>=10.0.0
moviepy==1.0.3
numpy>=1.24.0 